requirements.txt
python-socketio
Pillow
requests
websocket-client
numpy
opencv-python
pygame